fix(gig): include missing renames
