+++ libs/product-domain/backend/src/gig/domain/interfaces/worker-skills.ts
@@ -1,7 +1,7 @@
 import { WorkerSkill } from '@models/gig';
 import { CreateEntityInput } from '@models/shared';
 
-export type RegisterSkillInput = CreateEntityInput<
+export type RegisterWorkerSkillInput = CreateEntityInput<
   Omit<
     WorkerSkill,
     | 'experienceMonth'
