+++ libs/product-domain/backend/src/gig/domain/repositories/index.ts
@@ -4,7 +4,7 @@ export * from './gig-work';
 export * from './recommendation';
 export * from './review';
 export * from './skill-hire';
-export * from './skills';
+export * from './worker-skills';
 export * from './slot';
 export * from './update-transaction';
 export * from './worker';
